Add files via upload
The report html textbox input object was modified to include a title to allow the user to hover and see the entire url. The background JavaScript file was also modified to check for nulls and the code that was making the url on the input box turn red was removed.